fix : 일부 함수 및 컴포넌트 정리
