+++ src/features/problems/components/ImageCaptureInput.tsx
