+++ src/features/problems/components/ImageUploadSection.tsx
@@ -1,11 +1,18 @@
-import { Link } from 'react-router-dom';
-import EditChildInfoForm from '../../users/components/EditChildInfo';
 import { Camera } from 'lucide-react';
 import { useEffect, useRef, useState } from 'react';
+import ImageInputModal from './ImageInputModal';
 
 export default function ImageUploadSection() {
-  var divRef = useRef<HTMLDivElement>(null);
   const [height, setHeight] = useState<number>();
+  const [dialogOpen, setDialogOpen] = useState<boolean>(false);
+
+  var divRef = useRef<HTMLDivElement>(null);
+
+  const ClickEventHandler = () => {
+    if (!dialogOpen) {
+      setDialogOpen(true);
+    }
+  };
 
   useEffect(() => {
     const resizeObserver: ResizeObserver = new ResizeObserver(() => {
@@ -23,26 +30,19 @@ export default function ImageUploadSection() {
       resizeObserver.disconnect();
     };
   }, []);
+
   return (
-    <div className='w-full flex flex-col flex-grow'>
-      <div className='flex flex-row h-[35px] font-korean-title font-bold text-2xl gap-[7px] mb-3'>
-        <h2>안녕하세요</h2>
-        <div className='flex flex-row cursor-pointer'>
-          <div className='flex justify-center'>
-            <EditChildInfoForm />
-          </div>
-        </div>
-        <h2>부모님!</h2>
-      </div>
-      <Link
-        to='/problem/upload'
-        className='text-gray5 flex flex-col flex-grow items-center max-h-[40vh] max-w-full rounded-[24px] dark:text-gray2 inset-shadow-primary'
+    <>
+      <div
+        onClick={ClickEventHandler}
+        className='text-gray5 flex flex-col flex-grow items-center max-h-[40vh] max-w-full rounded-[24px] dark:text-gray2 inset-shadow-primary cursor-pointer'
       >
         <div ref={divRef} className='flex justify-center flex-grow flex-col'>
           <Camera width={120} height={height} />
           <span className='font-medium text-md'>문제를 등록해보세요</span>
         </div>
-      </Link>
-    </div>
+      </div>
+      <ImageInputModal dialogOpen={dialogOpen} setDialogOpen={setDialogOpen} />
+    </>
   );
 }
