+++ src/features/users/components/EditChildInfo.tsx
@@ -30,45 +30,57 @@ export default function EditChildInfoForm() {
   };
 
   return (
-    <Dialog>
-      <DialogTrigger asChild>
-        <Button size='sm_link' variant='link'>
-          <h2 className='text-primary'>{childName}</h2>
-          <SquarePen color='#ff7710' size={16.25} className='-ml-1 -mt-5' />
-        </Button>
-      </DialogTrigger>
-      <DialogContent className='flex flex-col items-center gap-6'>
-        <DialogHeader>
-          <DialogTitle>자녀 정보 수정</DialogTitle>
-        </DialogHeader>
-        <div className='w-full flex flex-col items-center gap-2'>
-          <Input
-            className='w-3/4 border-primary'
-            placeholder='이름을 입력하세요'
-            defaultValue={childName}
-            onChange={(e) => setTempName(e.target.value)}
-          />
-          <Select value={childGrade} onValueChange={setChildGrade}>
-            <SelectTrigger className='w-3/4'>
-              <SelectValue placeholder='학년을 선택하세요' />
-            </SelectTrigger>
-            <SelectContent>
-              {GRADE_OPTIONS.map(({ label, value }) => (
-                <SelectItem key={value} value={value}>
-                  {label}
-                </SelectItem>
-              ))}
-            </SelectContent>
-          </Select>
+    <div className='flex flex-row h-[35px] font-korean-title font-bold text-2xl gap-[7px] mb-3'>
+      <h2>안녕하세요</h2>
+      <div className='flex flex-row cursor-pointer'>
+        <div className='flex justify-center'>
+          <Dialog>
+            <DialogTrigger asChild>
+              <Button size='sm_link' variant='link'>
+                <h2 className='text-primary'>{childName}</h2>
+                <SquarePen
+                  color='#ff7710'
+                  size={16.25}
+                  className='-ml-1 -mt-5'
+                />
+              </Button>
+            </DialogTrigger>
+            <DialogContent className='flex flex-col items-center gap-6'>
+              <DialogHeader>
+                <DialogTitle>자녀 정보 수정</DialogTitle>
+              </DialogHeader>
+              <div className='w-full flex flex-col items-center gap-2'>
+                <Input
+                  className='w-3/4 border-primary'
+                  placeholder='이름을 입력하세요'
+                  defaultValue={childName}
+                  onChange={(e) => setTempName(e.target.value)}
+                />
+                <Select value={childGrade} onValueChange={setChildGrade}>
+                  <SelectTrigger className='w-3/4'>
+                    <SelectValue placeholder='학년을 선택하세요' />
+                  </SelectTrigger>
+                  <SelectContent>
+                    {GRADE_OPTIONS.map(({ label, value }) => (
+                      <SelectItem key={value} value={value}>
+                        {label}
+                      </SelectItem>
+                    ))}
+                  </SelectContent>
+                </Select>
+              </div>
+              <DialogFooter>
+                <DialogClose asChild>
+                  <Button onClick={clickHandler} size='lg' type='submit'>
+                    수정하기
+                  </Button>
+                </DialogClose>
+              </DialogFooter>
+            </DialogContent>
+          </Dialog>
         </div>
-        <DialogFooter>
-          <DialogClose asChild>
-            <Button onClick={clickHandler} size='lg' type='submit'>
-              수정하기
-            </Button>
-          </DialogClose>
-        </DialogFooter>
-      </DialogContent>
-    </Dialog>
+      </div>
+      <h2>부모님!</h2>
+    </div>
   );
 }
